Refactored out all storage account and config access to a new class called ConfigManager
