+++ NLog.Extensions.AzureTableStorage/NLog.Extensions.AzureTableStorage.csproj
@@ -76,6 +76,7 @@
   <ItemGroup>
     <Compile Include="AzureStorageTableNameValidator.cs" />
     <Compile Include="AzureTableStorageTarget.cs" />
+    <Compile Include="ConfigManager.cs" />
     <Compile Include="LogEntity.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TableStorageManager.cs" />
